+++ c3smembership/accountants_views.py
@@ -17,6 +17,7 @@ from c3smembership.models import (
     C3sMember,
     C3sStaff,
     # DBSession,
+    Dues15Invoice,
 )
 from c3smembership.utils import generate_pdf
 from c3smembership.mail_utils import (
@@ -48,7 +49,10 @@ from pyramid_mailer import get_mailer
 from pyramid_mailer.message import Message
 from pyramid.url import route_url
 from translationstring import TranslationStringFactory
-from datetime import datetime
+from datetime import (
+    datetime,
+    date,
+)
 # from sqlalchemy.exc import (
 #    IntegrityError,
 #    ResourceClosedError,
@@ -332,12 +336,20 @@ def switch_sig(request):
         )
     )
 
-    return HTTPFound(
-        request.route_url(
-            'dashboard',
-            number=dashboard_page, order=order, orderby=order_by)
-        + '#member_' + str(_member.id)
-    )
+    if 'dashboard' in request.referrer:
+        return HTTPFound(
+            request.route_url(
+                'dashboard',
+                number=dashboard_page, order=order, orderby=order_by)
+            + '#member_' + str(_member.id)
+        )
+    else:
+        return HTTPFound(
+            request.route_url(
+                'detail',
+                memberid=_member.id)
+            + '#membership_info'
+        )
 
 
 @view_config(permission='manage',
@@ -415,12 +427,20 @@ def switch_pay(request):
             _member.payment_received
         )
     )
-    return HTTPFound(
-        request.route_url(
-            'dashboard',
-            number=dashboard_page, order=order, orderby=order_by
-        ) + '#member_' + str(_member.id)
-    )
+    if 'dashboard' in request.referrer:
+        return HTTPFound(
+            request.route_url(
+                'dashboard',
+                number=dashboard_page, order=order, orderby=order_by)
+            + '#member_' + str(_member.id)
+        )
+    else:
+        return HTTPFound(
+            request.route_url(
+                'detail',
+                memberid=_member.id)
+            + '#membership_info'
+        )
 
 
 @view_config(renderer='json',
@@ -458,6 +478,8 @@ def member_detail(request):
     Mostly all the info about an application or membership
     in the database can be seen here.
     """
+    from decimal import Decimal as D
+    # import decimal
     logged_in = authenticated_userid(request)
 
     memberid = request.matchdict['memberid']
@@ -569,9 +591,15 @@ def member_detail(request):
         u'ü', u'ue').replace(
         u'.', u'')
 
+    # get the members invoices from the DB
+    _invoices = Dues15Invoice.get_by_membership_no(_member.membership_number)
+
     return {
+        'today': date.today().strftime('%Y-%m-%d'),
+        'D': D,
         'member': _member,
         'cert_link': _cert_link,
+        'invoices': _invoices,
         # 'form': html
     }